+++ src/api/auth.js
@@ -1,7 +1,9 @@
 import axiosInstance from './instance';
 
-export const signUp = (email, password) =>
-  axiosInstance.post('/auth/signup', { email, password });
+export const signUp = async (user) => {
+  return axiosInstance.post('/auth/signup', { ...user });
+};
 
-export const signIn = (email, password) =>
-  axiosInstance.post('/auth/signin', { email, password });
+export const signIn = async (user) => {
+  return axiosInstance.post('/auth/signin', { ...user });
+};
