--- src/api/core/api.js
@@ -1,15 +1,0 @@
-import axios from 'axios';
-
-export const client = axios.create({
-  baseURL: process.env.REACT_APP_API_URL,
-  headers: { 'Content-Type': 'application/json' },
-});
-
-client.interceptors.request.use((config) => {
-  const accessToken = localStorage.getItem('accessToken');
-
-  if (accessToken !== null) {
-    config.headers.Authorization = `Bearer ${accessToken}`;
-  }
-  return config;
-});
