+++ src/api/instance.js
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import { ACCESS_TOKEN } from '../constant/authConstant';
 
 const axiosConfig = {
   timeout: 3000,
@@ -12,7 +13,9 @@ axiosInstance.interceptors.request.use(
     if (!config.headers) config.headers = {};
     config.headers['Content-Type'] = 'application/json';
     config.headers['X-Requested-With'] = 'XMLHttpRequest';
-    config.headers.Authorization = `Bearer ${localStorage.getItem('token')}`;
+    config.headers.Authorization = `Bearer ${localStorage.getItem(
+      ACCESS_TOKEN,
+    )}`;
     config.headers.Accept = 'application/json';
     return config;
   },
@@ -23,6 +26,7 @@ axiosInstance.interceptors.response.use(
   (response) => response,
   (error) => {
     const { response } = error;
+    if (response.status === 401) return Promise.reject(error);
     if (response) alert(response.data.message);
     return Promise.reject(error);
   },
