+++ src/constant/authConstant.js
@@ -1,0 +1,1 @@
+export const ACCESS_TOKEN = 'accessToken';
